pyproject.toml
[tool.poetry]
name = "aseprite-websockets-mock"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "aseprite_websockets_mock"}]

[tool.poetry.dependencies]
python = "^3.8,<3.11"
fastapi = "^0.89.1"
websockets = "^10.4"
uvicorn = "^0.20.0"
Pillow = "^9.4.0"
httpx = "^0.23.3"
pytest-asyncio = "^0.20.3"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pytest = "^7.2.1"
websocket-client = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''
